Add test to install module
