+++ .github/workflows/ui-test/pr-ui.json
@@ -1,90 +1,214 @@
 {
   "include": [
     {
-      "comment": "1.7.7.8 PHP 7.3 online"
+      "comment": "1.7.0.6 PHP 7.1 online"
     },
     {
-      "PS_VERSION_START": "1.7.7.8",
-      "PHP_VERSION": "7.3",
+      "PS_VERSION_START": "1.7.0.6",
+      "PHP_VERSION": "7.1",
       "UPGRADE_CHANNEL": "online"
     },
     {
-      "comment": "8.0.3 PHP 7.3 online"
+      "comment": "1.7.1.2 PHP 7.1 online"
     },
     {
-      "PS_VERSION_START": "8.0.3",
-      "PHP_VERSION": "7.3",
+      "PS_VERSION_START": "1.7.1.2",
+      "PHP_VERSION": "7.1",
       "UPGRADE_CHANNEL": "online"
     },
     {
-      "comment": "8.0.3 PHP 8.1 online"
+      "comment": "1.7.2.5 PHP 7.1 online"
     },
     {
-      "PS_VERSION_START": "8.0.3",
-      "PHP_VERSION": "8.1",
+      "PS_VERSION_START": "1.7.2.5",
+      "PHP_VERSION": "7.1",
       "UPGRADE_CHANNEL": "online"
     },
     {
-      "comment": "8.2.1 PHP 8.1 online"
+      "comment": "1.7.3.4 PHP 7.1 online"
     },
     {
-      "PS_VERSION_START": "8.2.1",
-      "PHP_VERSION": "8.1",
+      "PS_VERSION_START": "1.7.3.4",
+      "PHP_VERSION": "7.1",
       "UPGRADE_CHANNEL": "online"
     },
     {
-      "comment": "8.2.3 PHP 7.2 ~ 8.1 online"
+      "comment": "1.7.5.2 PHP 7.1 online"
     },
     {
-      "PS_VERSION_START": "8.2.3",
+      "PS_VERSION_START": "1.7.5.2",
+      "PHP_VERSION": "7.1",
+      "UPGRADE_CHANNEL": "online"
+    },
+    {
+      "comment": "1.7.5.2 PHP 7.2 online"
+    },
+    {
+      "PS_VERSION_START": "1.7.5.2",
       "PHP_VERSION": "7.2",
       "UPGRADE_CHANNEL": "online"
     },
     {
-      "PS_VERSION_START": "8.2.3",
+      "comment": "1.7.6.9 PHP 7.1 online"
+    },
+    {
+      "PS_VERSION_START": "1.7.6.9",
+      "PHP_VERSION": "7.1",
+      "UPGRADE_CHANNEL": "online"
+    },
+    {
+      "comment": "1.7.6.9 PHP 7.2 online"
+    },
+    {
+      "PS_VERSION_START": "1.7.6.9",
+      "PHP_VERSION": "7.2",
+      "UPGRADE_CHANNEL": "online"
+    },
+    {
+      "comment": "1.7.7.8 PHP 7.1 online"
+    },
+    {
+      "PS_VERSION_START": "1.7.7.8",
+      "PHP_VERSION": "7.1",
+      "UPGRADE_CHANNEL": "online"
+    },
+    {
+      "comment": "1.7.7.8 PHP 7.2 ~ 7.3 online"
+    },
+    {
+      "PS_VERSION_START": "1.7.7.8",
+      "PHP_VERSION": "7.2",
+      "UPGRADE_CHANNEL": "online"
+    },
+    {
+      "PS_VERSION_START": "1.7.7.8",
       "PHP_VERSION": "7.3",
       "UPGRADE_CHANNEL": "online"
     },
     {
-      "PS_VERSION_START": "8.2.3",
+      "comment": "1.7.8.10 PHP 7.1 online"
+    },
+    {
+      "PS_VERSION_START": "1.7.8.10",
+      "PHP_VERSION": "7.1",
+      "UPGRADE_CHANNEL": "online"
+    },
+    {
+      "comment": "1.7.8.11 PHP 7.2 ~ 7.4 online"
+    },
+    {
+      "PS_VERSION_START": "1.7.8.11",
+      "PHP_VERSION": "7.2",
+      "UPGRADE_CHANNEL": "online"
+    },
+    {
+      "PS_VERSION_START": "1.7.8.11",
+      "PHP_VERSION": "7.3",
+      "UPGRADE_CHANNEL": "online"
+    },
+    {
+      "PS_VERSION_START": "1.7.8.11",
       "PHP_VERSION": "7.4",
       "UPGRADE_CHANNEL": "online"
     },
     {
-      "PS_VERSION_START": "8.2.3",
+      "comment": "8.0.5 PHP 7.2 ~ 8.1 online"
+    },
+    {
+      "PS_VERSION_START": "8.0.5",
+      "PHP_VERSION": "7.2",
+      "UPGRADE_CHANNEL": "online"
+    },
+    {
+      "PS_VERSION_START": "8.0.5",
+      "PHP_VERSION": "7.3",
+      "UPGRADE_CHANNEL": "online"
+    },
+    {
+      "PS_VERSION_START": "8.0.5",
+      "PHP_VERSION": "7.4",
+      "UPGRADE_CHANNEL": "online"
+    },
+    {
+      "PS_VERSION_START": "8.0.5",
       "PHP_VERSION": "8.0",
       "UPGRADE_CHANNEL": "online"
     },
     {
+      "PS_VERSION_START": "8.0.5",
+      "PHP_VERSION": "8.1",
+      "UPGRADE_CHANNEL": "online"
+    },
+    {
+      "comment": "8.1.7 PHP 7.2 ~ 8.1 online"
+    },
+    {
+      "PS_VERSION_START": "8.1.7",
+      "PHP_VERSION": "7.2",
+      "UPGRADE_CHANNEL": "online"
+    },
+    {
+      "PS_VERSION_START": "8.1.7",
+      "PHP_VERSION": "7.3",
+      "UPGRADE_CHANNEL": "online"
+    },
+    {
+      "PS_VERSION_START": "8.1.7",
+      "PHP_VERSION": "7.4",
+      "UPGRADE_CHANNEL": "online"
+    },
+    {
+      "PS_VERSION_START": "8.1.7",
+      "PHP_VERSION": "8.0",
+      "UPGRADE_CHANNEL": "online"
+    },
+    {
+      "PS_VERSION_START": "8.1.7",
+      "PHP_VERSION": "8.1",
+      "UPGRADE_CHANNEL": "online"
+    },
+    {
+      "comment": "8.2.3 PHP 8.1 online"
+    },
+    {
       "PS_VERSION_START": "8.2.3",
       "PHP_VERSION": "8.1",
       "UPGRADE_CHANNEL": "online"
     },
     {
-      "comment": "8.0.3 PHP 8.1 local"
+      "comment": "8.0.5 PHP 8.1 local"
     },
     {
-      "PS_VERSION_START": "8.0.3",
+      "PS_VERSION_START": "8.0.5",
       "PS_VERSION_END": "9.0.0",
       "PHP_VERSION": "8.1",
-      "UPGRADE_CHANNEL":  "local"
-    },
-    {
-      "comment": "8.0.3 PHP 7.3 local"
-    },
-    {
-      "PS_VERSION_START": "8.0.3",
-      "PS_VERSION_END": "8.1.0",
-      "PHP_VERSION": "7.3",
       "UPGRADE_CHANNEL": "local"
     },
-   {
-      "comment": "1.7.7.8 PHP 7.3 local"
-   },
-   {
-      "PS_VERSION_START": "1.7.7.8",
-      "PS_VERSION_END": "8.0.0",
-      "PHP_VERSION": "7.3",
+    {
+      "comment": "8.0.5 PHP 8.1 local"
+    },
+    {
+      "PS_VERSION_START": "8.0.5",
+      "PS_VERSION_END": "8.2.0",
+      "PHP_VERSION": "8.1",
+      "UPGRADE_CHANNEL": "local"
+    },
+    {
+      "comment": "8.1.7 PHP 8.1 local"
+    },
+    {
+      "PS_VERSION_START": "8.1.7",
+      "PS_VERSION_END": "9.0.0",
+      "PHP_VERSION": "8.1",
+      "UPGRADE_CHANNEL": "local"
+    },
+    {
+      "comment": "8.2.1 PHP 8.1 local"
+    },
+    {
+      "PS_VERSION_START": "8.2.1",
+      "PS_VERSION_END": "9.0.0",
+      "PHP_VERSION": "8.1",
       "UPGRADE_CHANNEL": "local"
     }
   ]
